Downloads Databrary contants from API.

Downloads Databrary contants from API.

assign_constants(vb = FALSE)

Arguments

vb
A Boolean value. If TRUE provides verbose output.

Value

A data frame with the constants.

Examples

assign_constants()

#> $permission
#> [1] "NONE"   "PUBLIC" "SHARED" "READ"   "EDIT"   "ADMIN" 
#> 
#> $release
#> [1] "PRIVATE"  "SHARED"   "EXCERPTS" "PUBLIC"  
#> 
#> $metric
#> $metric$`1`
#> $metric$`1`$id
#> [1] 1
#> 
#> $metric$`1`$category
#> [1] 1
#> 
#> $metric$`1`$name
#> [1] "ID"
#> 
#> $metric$`1`$release
#> [1] 3
#> 
#> $metric$`1`$type
#> [1] "text"
#> 
#> $metric$`1`$description
#> [1] "A unique, anonymized, primary identifier, such as participant ID"
#> 
#> $metric$`1`$required
#> [1] TRUE
#> 
#> 
#> $metric$`2`
#> $metric$`2`$id
#> [1] 2
#> 
#> $metric$`2`$category
#> [1] 1
#> 
#> $metric$`2`$name
#> [1] "info"
#> 
#> $metric$`2`$release
#> [1] 3
#> 
#> $metric$`2`$type
#> [1] "text"
#> 
#> $metric$`2`$description
#> [1] "Other information or alternate identifier"
#> 
#> 
#> $metric$`3`
#> $metric$`3`$id
#> [1] 3
#> 
#> $metric$`3`$category
#> [1] 1
#> 
#> $metric$`3`$name
#> [1] "description"
#> 
#> $metric$`3`$release
#> [1] 3
#> 
#> $metric$`3`$type
#> [1] "text"
#> 
#> $metric$`3`$long
#> [1] TRUE
#> 
#> $metric$`3`$description
#> [1] "A longer explanation or description"
#> 
#> 
#> $metric$`4`
#> $metric$`4`$id
#> [1] 4
#> 
#> $metric$`4`$category
#> [1] 1
#> 
#> $metric$`4`$name
#> [1] "birthdate"
#> 
#> $metric$`4`$type
#> [1] "date"
#> 
#> $metric$`4`$description
#> [1] "Date of birth (used with session date to calculate age; you can also use the group category to designate age groups)"
#> 
#> $metric$`4`$required
#> [1] FALSE
#> 
#> 
#> $metric$`5`
#> $metric$`5`$id
#> [1] 5
#> 
#> $metric$`5`$category
#> [1] 1
#> 
#> $metric$`5`$name
#> [1] "gender"
#> 
#> $metric$`5`$release
#> [1] 3
#> 
#> $metric$`5`$type
#> [1] "text"
#> 
#> $metric$`5`$options
#> [1] "Female" "Male"  
#> 
#> $metric$`5`$description
#> [1] "\"Male\", \"Female\", or any other relevant gender"
#> 
#> $metric$`5`$required
#> [1] FALSE
#> 
#> 
#> $metric$`6`
#> $metric$`6`$id
#> [1] 6
#> 
#> $metric$`6`$category
#> [1] 1
#> 
#> $metric$`6`$name
#> [1] "race"
#> 
#> $metric$`6`$release
#> [1] 3
#> 
#> $metric$`6`$type
#> [1] "text"
#> 
#> $metric$`6`$options
#> [1] "American Indian or Alaska Native"         
#> [2] "Asian"                                    
#> [3] "Native Hawaiian or Other Pacific Islander"
#> [4] "Black or African American"                
#> [5] "White"                                    
#> [6] "More than one"                            
#> [7] "Unknown or not reported"                  
#> 
#> $metric$`6`$description
#> [1] "As classified by NIH, or user-defined classification"
#> 
#> $metric$`6`$required
#> [1] FALSE
#> 
#> 
#> $metric$`7`
#> $metric$`7`$id
#> [1] 7
#> 
#> $metric$`7`$category
#> [1] 1
#> 
#> $metric$`7`$name
#> [1] "ethnicity"
#> 
#> $metric$`7`$release
#> [1] 3
#> 
#> $metric$`7`$type
#> [1] "text"
#> 
#> $metric$`7`$options
#> [1] "Not Hispanic or Latino"  "Hispanic or Latino"     
#> [3] "Unknown or not reported"
#> 
#> $metric$`7`$description
#> [1] "As classified by NIH (Hispanic/Non-Hispanic), or user-defined classification"
#> 
#> $metric$`7`$required
#> [1] FALSE
#> 
#> 
#> $metric$`8`
#> $metric$`8`$id
#> [1] 8
#> 
#> $metric$`8`$category
#> [1] 1
#> 
#> $metric$`8`$name
#> [1] "gestational age"
#> 
#> $metric$`8`$release
#> [1] 3
#> 
#> $metric$`8`$type
#> [1] "numeric"
#> 
#> $metric$`8`$description
#> [1] "Pregnancy age in weeks between last menstrual period and birth (or pre-natal observation)"
#> 
#> 
#> $metric$`9`
#> $metric$`9`$id
#> [1] 9
#> 
#> $metric$`9`$category
#> [1] 1
#> 
#> $metric$`9`$name
#> [1] "pregnancy term"
#> 
#> $metric$`9`$release
#> [1] 3
#> 
#> $metric$`9`$type
#> [1] "text"
#> 
#> $metric$`9`$options
#> [1] "Full term" "Preterm"  
#> 
#> $metric$`9`$assumed
#> [1] "Full term"
#> 
#> $metric$`9`$description
#> [1] "\"Full term\", \"Preterm\", or other gestational term"
#> 
#> 
#> $metric$`10`
#> $metric$`10`$id
#> [1] 10
#> 
#> $metric$`10`$category
#> [1] 1
#> 
#> $metric$`10`$name
#> [1] "birth weight"
#> 
#> $metric$`10`$release
#> [1] 3
#> 
#> $metric$`10`$type
#> [1] "numeric"
#> 
#> $metric$`10`$description
#> [1] "Weight at birth (in grams, e.g., 3250)"
#> 
#> 
#> $metric$`11`
#> $metric$`11`$id
#> [1] 11
#> 
#> $metric$`11`$category
#> [1] 1
#> 
#> $metric$`11`$name
#> [1] "disability"
#> 
#> $metric$`11`$type
#> [1] "text"
#> 
#> $metric$`11`$assumed
#> [1] "typical"
#> 
#> $metric$`11`$description
#> [1] "Any developmental, physical, or mental disability or disabilities"
#> 
#> $metric$`11`$required
#> [1] FALSE
#> 
#> 
#> $metric$`12`
#> $metric$`12`$id
#> [1] 12
#> 
#> $metric$`12`$category
#> [1] 1
#> 
#> $metric$`12`$name
#> [1] "language"
#> 
#> $metric$`12`$release
#> [1] 3
#> 
#> $metric$`12`$type
#> [1] "text"
#> 
#> $metric$`12`$assumed
#> [1] "English"
#> 
#> $metric$`12`$description
#> [1] "Primary language(s) spoken by and to participant"
#> 
#> $metric$`12`$required
#> [1] FALSE
#> 
#> 
#> $metric$`13`
#> $metric$`13`$id
#> [1] 13
#> 
#> $metric$`13`$category
#> [1] 1
#> 
#> $metric$`13`$name
#> [1] "country"
#> 
#> $metric$`13`$release
#> [1] 3
#> 
#> $metric$`13`$type
#> [1] "text"
#> 
#> $metric$`13`$assumed
#> [1] "US"
#> 
#> $metric$`13`$description
#> [1] "Country where participant was born"
#> 
#> 
#> $metric$`14`
#> $metric$`14`$id
#> [1] 14
#> 
#> $metric$`14`$category
#> [1] 1
#> 
#> $metric$`14`$name
#> [1] "state"
#> 
#> $metric$`14`$release
#> [1] 3
#> 
#> $metric$`14`$type
#> [1] "text"
#> 
#> $metric$`14`$options
#>  [1] "AL" "AK" "AZ" "AR" "CA" "CO" "CT" "DE" "DC" "FL" "GA" "HI" "ID" "IL" "IN"
#> [16] "IA" "KS" "KY" "LA" "ME" "MT" "NE" "NV" "NH" "NJ" "NM" "NY" "NC" "ND" "OH"
#> [31] "OK" "OR" "MD" "MA" "MI" "MN" "MS" "MO" "PA" "RI" "SC" "SD" "TN" "TX" "UT"
#> [46] "VT" "VA" "WA" "WV" "WI" "WY"
#> 
#> $metric$`14`$description
#> [1] "State/territory where participant was born"
#> 
#> 
#> $metric$`15`
#> $metric$`15`$id
#> [1] 15
#> 
#> $metric$`15`$category
#> [1] 1
#> 
#> $metric$`15`$name
#> [1] "setting"
#> 
#> $metric$`15`$release
#> [1] 3
#> 
#> $metric$`15`$type
#> [1] "text"
#> 
#> $metric$`15`$options
#> [1] "Lab"       "Home"      "Classroom" "Outdoor"   "Clinic"   
#> 
#> $metric$`15`$description
#> [1] "The physical context of the participant (please do not use for new data: see the context category instead)"
#> 
#> 
#> $metric$`16`
#> $metric$`16`$id
#> [1] 16
#> 
#> $metric$`16`$category
#> [1] 2
#> 
#> $metric$`16`$name
#> [1] "pilot"
#> 
#> $metric$`16`$release
#> [1] 3
#> 
#> $metric$`16`$type
#> [1] "void"
#> 
#> $metric$`16`$required
#> [1] FALSE
#> 
#> 
#> $metric$`17`
#> $metric$`17`$id
#> [1] 17
#> 
#> $metric$`17`$category
#> [1] 2
#> 
#> $metric$`17`$name
#> [1] "name"
#> 
#> $metric$`17`$release
#> [1] 3
#> 
#> $metric$`17`$type
#> [1] "text"
#> 
#> $metric$`17`$description
#> [1] "A label or identifier referring to the pilot method"
#> 
#> 
#> $metric$`18`
#> $metric$`18`$id
#> [1] 18
#> 
#> $metric$`18`$category
#> [1] 2
#> 
#> $metric$`18`$name
#> [1] "description"
#> 
#> $metric$`18`$release
#> [1] 3
#> 
#> $metric$`18`$type
#> [1] "text"
#> 
#> $metric$`18`$long
#> [1] TRUE
#> 
#> $metric$`18`$description
#> [1] "A longer explanation or description of the pilot method"
#> 
#> 
#> $metric$`19`
#> $metric$`19`$id
#> [1] 19
#> 
#> $metric$`19`$category
#> [1] 3
#> 
#> $metric$`19`$name
#> [1] "excluded"
#> 
#> $metric$`19`$release
#> [1] 3
#> 
#> $metric$`19`$type
#> [1] "void"
#> 
#> 
#> $metric$`20`
#> $metric$`20`$id
#> [1] 20
#> 
#> $metric$`20`$category
#> [1] 3
#> 
#> $metric$`20`$name
#> [1] "name"
#> 
#> $metric$`20`$release
#> [1] 3
#> 
#> $metric$`20`$type
#> [1] "text"
#> 
#> $metric$`20`$description
#> [1] "A label or identifier referring to the exclusion criterion"
#> 
#> 
#> $metric$`21`
#> $metric$`21`$id
#> [1] 21
#> 
#> $metric$`21`$category
#> [1] 3
#> 
#> $metric$`21`$name
#> [1] "reason"
#> 
#> $metric$`21`$release
#> [1] 3
#> 
#> $metric$`21`$type
#> [1] "text"
#> 
#> $metric$`21`$options
#> [1] "Did not meet inclusion criteria" "Procedural/experimenter error"  
#> [3] "Withdrew/fussy/tired"            "Outlier"                        
#> 
#> $metric$`21`$description
#> [1] "The reason for excluding these data"
#> 
#> $metric$`21`$required
#> [1] FALSE
#> 
#> 
#> $metric$`22`
#> $metric$`22`$id
#> [1] 22
#> 
#> $metric$`22`$category
#> [1] 3
#> 
#> $metric$`22`$name
#> [1] "description"
#> 
#> $metric$`22`$release
#> [1] 3
#> 
#> $metric$`22`$type
#> [1] "text"
#> 
#> $metric$`22`$long
#> [1] TRUE
#> 
#> $metric$`22`$description
#> [1] "A longer explanation or description of the reason for excluding data"
#> 
#> 
#> $metric$`23`
#> $metric$`23`$id
#> [1] 23
#> 
#> $metric$`23`$category
#> [1] 4
#> 
#> $metric$`23`$name
#> [1] "name"
#> 
#> $metric$`23`$release
#> [1] 3
#> 
#> $metric$`23`$type
#> [1] "text"
#> 
#> $metric$`23`$description
#> [1] "A label or identifier for the condition"
#> 
#> $metric$`23`$required
#> [1] TRUE
#> 
#> 
#> $metric$`24`
#> $metric$`24`$id
#> [1] 24
#> 
#> $metric$`24`$category
#> [1] 4
#> 
#> $metric$`24`$name
#> [1] "description"
#> 
#> $metric$`24`$release
#> [1] 3
#> 
#> $metric$`24`$type
#> [1] "text"
#> 
#> $metric$`24`$long
#> [1] TRUE
#> 
#> $metric$`24`$description
#> [1] "A longer explanation or description of the condition"
#> 
#> 
#> $metric$`25`
#> $metric$`25`$id
#> [1] 25
#> 
#> $metric$`25`$category
#> [1] 4
#> 
#> $metric$`25`$name
#> [1] "info"
#> 
#> $metric$`25`$release
#> [1] 3
#> 
#> $metric$`25`$type
#> [1] "text"
#> 
#> $metric$`25`$description
#> [1] "Other information or alternate identifier"
#> 
#> 
#> $metric$`26`
#> $metric$`26`$id
#> [1] 26
#> 
#> $metric$`26`$category
#> [1] 5
#> 
#> $metric$`26`$name
#> [1] "name"
#> 
#> $metric$`26`$release
#> [1] 3
#> 
#> $metric$`26`$type
#> [1] "text"
#> 
#> $metric$`26`$description
#> [1] "A label or identifier for the grouping"
#> 
#> $metric$`26`$required
#> [1] TRUE
#> 
#> 
#> $metric$`27`
#> $metric$`27`$id
#> [1] 27
#> 
#> $metric$`27`$category
#> [1] 5
#> 
#> $metric$`27`$name
#> [1] "description"
#> 
#> $metric$`27`$release
#> [1] 3
#> 
#> $metric$`27`$type
#> [1] "text"
#> 
#> $metric$`27`$long
#> [1] TRUE
#> 
#> $metric$`27`$description
#> [1] "A longer explanation or description of the grouping"
#> 
#> 
#> $metric$`28`
#> $metric$`28`$id
#> [1] 28
#> 
#> $metric$`28`$category
#> [1] 5
#> 
#> $metric$`28`$name
#> [1] "info"
#> 
#> $metric$`28`$release
#> [1] 3
#> 
#> $metric$`28`$type
#> [1] "text"
#> 
#> $metric$`28`$description
#> [1] "Other information or alternate identifier"
#> 
#> 
#> $metric$`29`
#> $metric$`29`$id
#> [1] 29
#> 
#> $metric$`29`$category
#> [1] 6
#> 
#> $metric$`29`$name
#> [1] "name"
#> 
#> $metric$`29`$release
#> [1] 3
#> 
#> $metric$`29`$type
#> [1] "text"
#> 
#> $metric$`29`$description
#> [1] "A label or identifier for the task"
#> 
#> $metric$`29`$required
#> [1] TRUE
#> 
#> 
#> $metric$`30`
#> $metric$`30`$id
#> [1] 30
#> 
#> $metric$`30`$category
#> [1] 6
#> 
#> $metric$`30`$name
#> [1] "description"
#> 
#> $metric$`30`$release
#> [1] 3
#> 
#> $metric$`30`$type
#> [1] "text"
#> 
#> $metric$`30`$long
#> [1] TRUE
#> 
#> $metric$`30`$description
#> [1] "A longer explanation or description of the task"
#> 
#> $metric$`30`$required
#> [1] FALSE
#> 
#> 
#> $metric$`31`
#> $metric$`31`$id
#> [1] 31
#> 
#> $metric$`31`$category
#> [1] 6
#> 
#> $metric$`31`$name
#> [1] "info"
#> 
#> $metric$`31`$release
#> [1] 3
#> 
#> $metric$`31`$type
#> [1] "text"
#> 
#> $metric$`31`$description
#> [1] "Other information or alternate identifier"
#> 
#> 
#> $metric$`32`
#> $metric$`32`$id
#> [1] 32
#> 
#> $metric$`32`$category
#> [1] 7
#> 
#> $metric$`32`$name
#> [1] "name"
#> 
#> $metric$`32`$release
#> [1] 3
#> 
#> $metric$`32`$type
#> [1] "text"
#> 
#> $metric$`32`$description
#> [1] "A label or identifier for the context"
#> 
#> 
#> $metric$`33`
#> $metric$`33`$id
#> [1] 33
#> 
#> $metric$`33`$category
#> [1] 7
#> 
#> $metric$`33`$name
#> [1] "setting"
#> 
#> $metric$`33`$release
#> [1] 3
#> 
#> $metric$`33`$type
#> [1] "text"
#> 
#> $metric$`33`$options
#> [1] "Lab"       "Home"      "Classroom" "Outdoor"   "Clinic"   
#> 
#> $metric$`33`$description
#> [1] "The physical context"
#> 
#> $metric$`33`$required
#> [1] TRUE
#> 
#> 
#> $metric$`34`
#> $metric$`34`$id
#> [1] 34
#> 
#> $metric$`34`$category
#> [1] 7
#> 
#> $metric$`34`$name
#> [1] "language"
#> 
#> $metric$`34`$release
#> [1] 3
#> 
#> $metric$`34`$type
#> [1] "text"
#> 
#> $metric$`34`$assumed
#> [1] "English"
#> 
#> $metric$`34`$description
#> [1] "Language used in this context"
#> 
#> $metric$`34`$required
#> [1] FALSE
#> 
#> 
#> $metric$`35`
#> $metric$`35`$id
#> [1] 35
#> 
#> $metric$`35`$category
#> [1] 7
#> 
#> $metric$`35`$name
#> [1] "country"
#> 
#> $metric$`35`$release
#> [1] 3
#> 
#> $metric$`35`$type
#> [1] "text"
#> 
#> $metric$`35`$assumed
#> [1] "US"
#> 
#> $metric$`35`$description
#> [1] "Country of data collection"
#> 
#> $metric$`35`$required
#> [1] FALSE
#> 
#> 
#> $metric$`36`
#> $metric$`36`$id
#> [1] 36
#> 
#> $metric$`36`$category
#> [1] 7
#> 
#> $metric$`36`$name
#> [1] "state"
#> 
#> $metric$`36`$release
#> [1] 3
#> 
#> $metric$`36`$type
#> [1] "text"
#> 
#> $metric$`36`$options
#>  [1] "AL" "AK" "AZ" "AR" "CA" "CO" "CT" "DE" "DC" "FL" "GA" "HI" "ID" "IL" "IN"
#> [16] "IA" "KS" "KY" "LA" "ME" "MT" "NE" "NV" "NH" "NJ" "NM" "NY" "NC" "ND" "OH"
#> [31] "OK" "OR" "MD" "MA" "MI" "MN" "MS" "MO" "PA" "RI" "SC" "SD" "TN" "TX" "UT"
#> [46] "VT" "VA" "WA" "WV" "WI" "WY"
#> 
#> $metric$`36`$description
#> [1] "State/territory of data collection"
#> 
#> $metric$`36`$required
#> [1] FALSE
#> 
#> 
#> 
#> $category
#> $category$`1`
#> $category$`1`$id
#> [1] 1
#> 
#> $category$`1`$name
#> [1] "participant"
#> 
#> $category$`1`$description
#> [1] "An individual human subject whose data are used or represented"
#> 
#> 
#> $category$`2`
#> $category$`2`$id
#> [1] 2
#> 
#> $category$`2`$name
#> [1] "pilot"
#> 
#> $category$`2`$description
#> [1] "Indicates that the methods used were not finalized or were non-standard"
#> 
#> 
#> $category$`3`
#> $category$`3`$id
#> [1] 3
#> 
#> $category$`3`$name
#> [1] "exclusion"
#> 
#> $category$`3`$description
#> [1] "Indicates that data were not usable"
#> 
#> 
#> $category$`4`
#> $category$`4`$id
#> [1] 4
#> 
#> $category$`4`$name
#> [1] "condition"
#> 
#> $category$`4`$description
#> [1] "An experimenter-determined manipulation (within or between sessions)"
#> 
#> 
#> $category$`5`
#> $category$`5`$id
#> [1] 5
#> 
#> $category$`5`$name
#> [1] "group"
#> 
#> $category$`5`$description
#> [1] "A grouping determined by an aspect of the data (participant ability, age, grade level, experience, longitudinal visit, measurements used/available)"
#> 
#> 
#> $category$`6`
#> $category$`6`$id
#> [1] 6
#> 
#> $category$`6`$name
#> [1] "task"
#> 
#> $category$`6`$description
#> [1] "A particular task, activity, or phase of the session or study"
#> 
#> 
#> $category$`7`
#> $category$`7`$id
#> [1] 7
#> 
#> $category$`7`$name
#> [1] "context"
#> 
#> $category$`7`$description
#> [1] "A particular setting or other aspect of where/when/how data were collected"
#> 
#> 
#> 
#> $format
#> $format$`-800`
#> $format$`-800`$id
#> [1] -800
#> 
#> $format$`-800`$mimetype
#> [1] "video/mp4"
#> 
#> $format$`-800`$extension
#> [1] "mp4"
#> 
#> $format$`-800`$name
#> [1] "MPEG-4 video"
#> 
#> $format$`-800`$transcodable
#> [1] -800
#> 
#> 
#> $format$`-700`
#> $format$`-700`$id
#> [1] -700
#> 
#> $format$`-700`$mimetype
#> [1] "image/jpeg"
#> 
#> $format$`-700`$extension
#> [1] "jpg"
#> 
#> $format$`-700`$name
#> [1] "JPEG image"
#> 
#> 
#> $format$`-600`
#> $format$`-600`$id
#> [1] -600
#> 
#> $format$`-600`$mimetype
#> [1] "audio/mpeg"
#> 
#> $format$`-600`$extension
#> [1] "mp3"
#> 
#> $format$`-600`$name
#> [1] "MPEG-1 or MPEG-2 audio layer III"
#> 
#> $format$`-600`$transcodable
#> [1] -600
#> 
#> 
#> $format$`1`
#> $format$`1`$id
#> [1] 1
#> 
#> $format$`1`$mimetype
#> [1] "text/plain"
#> 
#> $format$`1`$extension
#> [1] "txt"
#> 
#> $format$`1`$name
#> [1] "Plain text"
#> 
#> 
#> $format$`2`
#> $format$`2`$id
#> [1] 2
#> 
#> $format$`2`$mimetype
#> [1] "text/csv"
#> 
#> $format$`2`$extension
#> [1] "csv"
#> 
#> $format$`2`$name
#> [1] "Comma-separated values"
#> 
#> 
#> $format$`4`
#> $format$`4`$id
#> [1] 4
#> 
#> $format$`4`$mimetype
#> [1] "text/rtf"
#> 
#> $format$`4`$extension
#> [1] "rtf"
#> 
#> $format$`4`$name
#> [1] "Rich text format"
#> 
#> 
#> $format$`5`
#> $format$`5`$id
#> [1] 5
#> 
#> $format$`5`$mimetype
#> [1] "image/png"
#> 
#> $format$`5`$extension
#> [1] "png"
#> 
#> $format$`5`$name
#> [1] "Portable network graphics"
#> 
#> 
#> $format$`6`
#> $format$`6`$id
#> [1] 6
#> 
#> $format$`6`$mimetype
#> [1] "application/pdf"
#> 
#> $format$`6`$extension
#> [1] "pdf"
#> 
#> $format$`6`$name
#> [1] "Portable document"
#> 
#> 
#> $format$`7`
#> $format$`7`$id
#> [1] 7
#> 
#> $format$`7`$mimetype
#> [1] "application/msword"
#> 
#> $format$`7`$extension
#> [1] "doc"
#> 
#> $format$`7`$name
#> [1] "Microsoft Word document"
#> 
#> 
#> $format$`8`
#> $format$`8`$id
#> [1] 8
#> 
#> $format$`8`$mimetype
#> [1] "application/vnd.oasis.opendocument.text"
#> 
#> $format$`8`$extension
#> [1] "odf"
#> 
#> $format$`8`$name
#> [1] "OpenDocument text"
#> 
#> 
#> $format$`9`
#> $format$`9`$id
#> [1] 9
#> 
#> $format$`9`$mimetype
#> [1] "application/vnd.openxmlformats-officedocument.wordprocessingml.document"
#> 
#> $format$`9`$extension
#> [1] "docx"
#> 
#> $format$`9`$name
#> [1] "Microsoft Word (Office Open XML) document"
#> 
#> 
#> $format$`10`
#> $format$`10`$id
#> [1] 10
#> 
#> $format$`10`$mimetype
#> [1] "application/vnd.ms-excel"
#> 
#> $format$`10`$extension
#> [1] "xls"
#> 
#> $format$`10`$name
#> [1] "Microsoft Excel spreadsheet"
#> 
#> 
#> $format$`11`
#> $format$`11`$id
#> [1] 11
#> 
#> $format$`11`$mimetype
#> [1] "application/vnd.oasis.opendocument.spreadsheet"
#> 
#> $format$`11`$extension
#> [1] "ods"
#> 
#> $format$`11`$name
#> [1] "OpenDocument spreadsheet"
#> 
#> 
#> $format$`12`
#> $format$`12`$id
#> [1] 12
#> 
#> $format$`12`$mimetype
#> [1] "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet"
#> 
#> $format$`12`$extension
#> [1] "xlsx"
#> 
#> $format$`12`$name
#> [1] "Microsoft Excel (Office Open XML) workbook"
#> 
#> 
#> $format$`13`
#> $format$`13`$id
#> [1] 13
#> 
#> $format$`13`$mimetype
#> [1] "application/vnd.ms-powerpoint"
#> 
#> $format$`13`$extension
#> [1] "ppt"
#> 
#> $format$`13`$name
#> [1] "Microsoft PowerPoint presentation"
#> 
#> 
#> $format$`14`
#> $format$`14`$id
#> [1] 14
#> 
#> $format$`14`$mimetype
#> [1] "application/vnd.oasis.opendocument.presentation"
#> 
#> $format$`14`$extension
#> [1] "odp"
#> 
#> $format$`14`$name
#> [1] "OpenDocument presentation"
#> 
#> 
#> $format$`15`
#> $format$`15`$id
#> [1] 15
#> 
#> $format$`15`$mimetype
#> [1] "application/vnd.openxmlformats-officedocument.presentationml.presentation"
#> 
#> $format$`15`$extension
#> [1] "pptx"
#> 
#> $format$`15`$name
#> [1] "Microsoft PowerPoint (Office Open XML) presentation"
#> 
#> 
#> $format$`16`
#> $format$`16`$id
#> [1] 16
#> 
#> $format$`16`$mimetype
#> [1] "application/vnd.datavyu"
#> 
#> $format$`16`$extension
#> [1] "opf"
#> 
#> $format$`16`$name
#> [1] "Datavyu"
#> 
#> 
#> $format$`18`
#> $format$`18`$id
#> [1] 18
#> 
#> $format$`18`$mimetype
#> [1] "video/webm"
#> 
#> $format$`18`$extension
#> [1] "webm"
#> 
#> $format$`18`$name
#> [1] "WebM video"
#> 
#> $format$`18`$transcodable
#> [1] -800
#> 
#> 
#> $format$`19`
#> $format$`19`$id
#> [1] 19
#> 
#> $format$`19`$mimetype
#> [1] "video/mpeg"
#> 
#> $format$`19`$extension
#> [1] "mpg"
#> 
#> $format$`19`$name
#> [1] "MPEG program stream (MPEG-1/MPEG-2 video)"
#> 
#> $format$`19`$transcodable
#> [1] -800
#> 
#> 
#> $format$`20`
#> $format$`20`$id
#> [1] 20
#> 
#> $format$`20`$mimetype
#> [1] "video/quicktime"
#> 
#> $format$`20`$extension
#> [1] "mov"
#> 
#> $format$`20`$name
#> [1] "QuickTime video"
#> 
#> $format$`20`$transcodable
#> [1] -800
#> 
#> 
#> $format$`21`
#> $format$`21`$id
#> [1] 21
#> 
#> $format$`21`$mimetype
#> [1] "video/mp2t"
#> 
#> $format$`21`$extension
#> [1] "mts"
#> 
#> $format$`21`$name
#> [1] "MPEG transport stream"
#> 
#> $format$`21`$transcodable
#> [1] -800
#> 
#> 
#> $format$`22`
#> $format$`22`$id
#> [1] 22
#> 
#> $format$`22`$mimetype
#> [1] "video/avi"
#> 
#> $format$`22`$extension
#> [1] "avi"
#> 
#> $format$`22`$name
#> [1] "Audio Video Interleave"
#> 
#> $format$`22`$transcodable
#> [1] -800
#> 
#> 
#> $format$`23`
#> $format$`23`$id
#> [1] 23
#> 
#> $format$`23`$mimetype
#> [1] "application/x-spss-sav"
#> 
#> $format$`23`$extension
#> [1] "sav"
#> 
#> $format$`23`$name
#> [1] "SPSS System File"
#> 
#> 
#> $format$`24`
#> $format$`24`$id
#> [1] 24
#> 
#> $format$`24`$mimetype
#> [1] "audio/wav"
#> 
#> $format$`24`$extension
#> [1] "wav"
#> 
#> $format$`24`$name
#> [1] "Waveform audio"
#> 
#> $format$`24`$transcodable
#> [1] -600
#> 
#> 
#> $format$`25`
#> $format$`25`$id
#> [1] 25
#> 
#> $format$`25`$mimetype
#> [1] "video/x-ms-wmv"
#> 
#> $format$`25`$extension
#> [1] "wmv"
#> 
#> $format$`25`$name
#> [1] "Windows Media video"
#> 
#> $format$`25`$transcodable
#> [1] -800
#> 
#> 
#> $format$`26`
#> $format$`26`$id
#> [1] 26
#> 
#> $format$`26`$mimetype
#> [1] "text/x-chat"
#> 
#> $format$`26`$extension
#> [1] "cha"
#> 
#> $format$`26`$name
#> [1] "Codes for the Human Analysis of Transcripts"
#> 
#> 
#> $format$`27`
#> $format$`27`$id
#> [1] 27
#> 
#> $format$`27`$mimetype
#> [1] "audio/aac"
#> 
#> $format$`27`$extension
#> [1] "aac"
#> 
#> $format$`27`$name
#> [1] "Advanced Audio Coding"
#> 
#> $format$`27`$transcodable
#> [1] -600
#> 
#> 
#> $format$`28`
#> $format$`28`$id
#> [1] 28
#> 
#> $format$`28`$mimetype
#> [1] "audio/x-ms-wma"
#> 
#> $format$`28`$extension
#> [1] "wma"
#> 
#> $format$`28`$name
#> [1] "Windows Media audio"
#> 
#> $format$`28`$transcodable
#> [1] -600
#> 
#> 
#> $format$`29`
#> $format$`29`$id
#> [1] 29
#> 
#> $format$`29`$mimetype
#> [1] "application/vnd.lena.interpreted-time-segments"
#> 
#> $format$`29`$extension
#> [1] "its"
#> 
#> $format$`29`$name
#> [1] "LENA Interpreted Time Segments"
#> 
#> 
#> $format$`30`
#> $format$`30`$id
#> [1] 30
#> 
#> $format$`30`$mimetype
#> [1] "video/x-dv"
#> 
#> $format$`30`$extension
#> [1] "dv"
#> 
#> $format$`30`$name
#> [1] "Digital Interface Format video"
#> 
#> $format$`30`$transcodable
#> [1] -800
#> 
#> 
#> $format$`31`
#> $format$`31`$id
#> [1] 31
#> 
#> $format$`31`$mimetype
#> [1] "text/elan"
#> 
#> $format$`31`$extension
#> [1] "eaf"
#> 
#> $format$`31`$name
#> [1] "ELAN - Linguistic Annotator"
#> 
#> 
#> 
#> $party
#> $party$nobody
#> $party$nobody$id
#> [1] -1
#> 
#> $party$nobody$sortname
#> [1] "Everybody"
#> 
#> $party$nobody$institution
#> [1] TRUE
#> 
#> 
#> $party$root
#> $party$root$id
#> [1] 0
#> 
#> $party$root$sortname
#> [1] "Databrary"
#> 
#> $party$root$institution
#> [1] TRUE
#> 
#> 
#> $party$staff
#> $party$staff$id
#> [1] 2
#> 
#> $party$staff$sortname
#> [1] "Staff"
#> 
#> $party$staff$affiliation
#> [1] "Databrary"
#> 
#> $party$staff$institution
#> [1] TRUE
#> 
#> 
#> 
#> $notice
#> $notice$VolumeAccess
#> [1] 14
#> 
#> $notice$ReleaseExcerpt
#> [1] 17
#> 
#> $notice$AuthorizeChildExpiring
#> [1] 8
#> 
#> $notice$VolumeAssist
#> [1] 10
#> 
#> $notice$AuthorizeRequest
#> [1] 2
#> 
#> $notice$TagVolume
#> [1] 21
#> 
#> $notice$VolumeAccessOther
#> [1] 13
#> 
#> $notice$VolumeCreated
#> [1] 11
#> 
#> $notice$SharedVolume
#> [1] 22
#> 
#> $notice$Newsletter
#> [1] 23
#> 
#> $notice$CommentReply
#> [1] 20
#> 
#> $notice$AuthorizeGranted
#> [1] 3
#> 
#> $notice$AuthorizeExpired
#> [1] 5
#> 
#> $notice$AccountChange
#> [1] 1
#> 
#> $notice$ExcerptVolume
#> [1] 18
#> 
#> $notice$ReleaseSlot
#> [1] 15
#> 
#> $notice$AuthorizeExpiring
#> [1] 4
#> 
#> $notice$VolumeSharing
#> [1] 12
#> 
#> $notice$AuthorizeChildRequest
#> [1] 6
#> 
#> $notice$CommentVolume
#> [1] 19
#> 
#> $notice$AuthorizeChildExpired
#> [1] 9
#> 
#> $notice$AuthorizeChildGranted
#> [1] 7
#> 
#> $notice$ReleaseAsset
#> [1] 16
#> 
#> 
#> $delivery
#> [1] "none"   "site"   "weekly" "daily"  "async" 
#> 
#> $version
#> [1] "1"
#>