Downloads container and record structure from Databrary volume. This function is deprecated, as it duplicates list_containers_records().

Downloads container and record structure from Databrary volume. This function is deprecated, as it duplicates list_containers_records().

download_containers_records(vol_id = 1, vb = FALSE)

Arguments

vol_id
Databrary volume number.

vb
A boolean value. If TRUE provides verbose output.

Value

List of containers and records from the specified volume.

Examples

download_containers_records()

#> $id
#> [1] 1
#> 
#> $name
#> [1] "Databrary sponsored workshops and events"
#> 
#> $body
#> [1] "Our mission is to promote sharing and reuse of video data. Most developmental scientists rely on video recordings to capture the complexity and richness of behavior. However, researchers rarely share video data, and this has impeded scientific progress. Enabling reuse of video data is essential to increasing openness, maximizing the efficiency of limited resources, and accelerating the pace of discovery in the developmental and learning sciences.\n\nWe aim to provide tools that encourage good data management, facilitate open data sharing, enable discovery, and foster data reuse and integration. We are working to build functionality that allows researchers to organize, store, code, and share identifiable video data and other contextual information about the videos with other researchers through a secure web-based library.\n\nThe Databrary project is dedicated to transforming the culture of developmental science by building a community of researchers committed to open video data sharing, training a new generation of developmental scientists and empowering them with an unprecedented set of tools for discovery, and raising the profile of behavioral science by bolstering interest in and support for scientific research among the general public."
#> 
#> $doi
#> [1] "10.17910/B7159Q"
#> 
#> $creation
#> [1] "2013-01-11T10:26:40Z"
#> 
#> $owners
#>               name   id
#> 1    Adolph, Karen    5
#> 2 Gilmore, Rick O.    6
#> 3            Staff  308
#> 4 Admin, Databrary 3341
#> 
#> $permission
#> [1] 1
#> 
#> $publicsharefull
#> [1] TRUE
#> 
#> $publicaccess
#> [1] "full"
#> 
#> $containers
#>       id  top                                                      name
#> 1      1 TRUE                                                      <NA>
#> 2   6256   NA                                    Advisory Board Meeting
#> 3   6257   NA                                    Advisory Board Meeting
#> 4   6540   NA                                    Advisory Board Meeting
#> 5   7821   NA                               SRCD Preconference Workshop
#> 6   8460   NA                                    Advisory Board Meeting
#> 7   9224   NA                    Regional Workshop at NYU (Slides only)
#> 8   9578   NA                                CDS Preconference Workshop
#> 9   9807 TRUE                                       Top-level materials
#> 10 10068   NA                         Regional Workshop in Philadelphia
#> 11 10991   NA                           Regional Workshop at Penn State
#> 12 11362   NA Regional Workshop at University of Maryland, College Park
#> 13 11660   NA                             ICIS16 Preconference Workshop
#> 14 11886   NA                                    Advisory Board Meeting
#> 15 14297   NA              Regional Workshop at University of Minnesota
#> 16 14550   NA                              Regional Workshop at Harvard
#> 17 14551   NA                               Regional Workshop in Boston
#> 18 15663   NA                                    Bethesda PLAY Workshop
#> 19 16104   NA                   Regional Workshop at University of Iowa
#> 20 16891   NA                                    Pre-Conference at SRCD
#> 21 16894   NA Pre-Conference at SRCD "The Science of Play and Learning"
#> 22 17187   NA                                     AERA 2017 San Antonio
#> 23 17417   NA      Regional Workshop at Emory University in Atlanta, GA
#> 24 17886   NA                               Advisory Board Meeting 2017
#> 25 21077   NA                                         CDS 2017 Portland
#> 26 25952   NA                               Advisory Board Meeting 2018
#>          date release
#> 1        <NA>      NA
#> 2  2013-10-28       3
#> 3  2014-04-07       3
#> 4  2014-10-17       3
#> 5  2015-03-18       3
#> 6  2015-05-11       3
#> 7  2015-09-18       3
#> 8  2015-10-08       3
#> 9        <NA>      NA
#> 10 2015-11-20       3
#> 11 2016-03-25       3
#> 12 2016-04-15       3
#> 13 2016-05-25       3
#> 14 2016-06-13       3
#> 15 2016-10-14       3
#> 16 2016-10-21       3
#> 17 2016-10-21       3
#> 18 2016-12-16       3
#> 19 2017-02-03       3
#> 20 2017-04-05       3
#> 21 2017-04-05       3
#> 22 2017-04-30       3
#> 23 2017-05-15       3
#> 24 2017-06-09       3
#> 25 2017-10-13       3
#> 26 2018-03-26       3
#> 
#> $records
#>     id category                                        29
#> 1 5481        6 Datavyu & Best Practices for Coding Video
#> 2 5480        6              Managing Data with Databrary
#> 3 5479        6   Policies for Sharing Identifiable Video
#> 4 5478        6                 Introduction to Databrary
#>