Converts a Datavyu (.opf) file to CSV.

Converts a Datavyu (.opf) file to CSV.

dv_to_csv(
  dv_dir = ".",
  dv_fn = "db",
  out_fn = paste0(dv_dir, "/tmp.csv"),
  auto_write_over = FALSE,
  code_regex = "^([a-zA-Z_]+[0-9]*[a-zA-Z_]*[0-9]*)",
  code_type_regex = "([a-zA-Z]+)$",
  time_regex = "([0-9]{2}:[0-9]{2}:[0-9]{2}:[0-9]{3})",
  code_values_regex = "\\(([a-zA-Z ?,.'/0-9;!|~`]+)\\)$",
  convert_times = TRUE,
  vb = FALSE
)

Arguments

dv_dir

Directory to extracted Datavyu file.

dv_fn

Datavyu code file. Defaults to 'db'.

out_fn

Output file name. Default is based on dv_dir.

auto_write_over

A Boolean value. If TRUE, new output file overwrites old.

code_regex

Regular expression to extract codes from Datavyu file.

code_type_regex

Regular expression to extract doce types from Datavyu file.

time_regex

Regular expression to extract onset/offset times.

code_values_regex

Regular expression to extract code values from Datavyu file.

convert_times

A Boolean value. If TRUE (default), converts the onset and offset

vb

A boolean value. If TRUE, provides verbose output.

Examples

dv_to_csv()
#> Warning: cannot open file './db': No such file or directory
#> Error in file(paste0(dv_dir, "/", dv_fn), "r"): cannot open the connection