Lists data for a given Databrary party (institution or person).

Lists data for a given Databrary party (institution or person).

get_volume(vol_id = 1, search_str = "", vb = FALSE)

Arguments

vol_id
Target volume number.

search_str
The specific data to retrieve for the volume

vb
A Boolean value. If TRUE provides verbose output.

Value

A list (or data frame) with the requested data.

Examples

get_volume() # Default is volume 1

#> $id
#> [1] 1
#> 
#> $name
#> [1] "Databrary sponsored workshops and events"
#> 
#> $body
#> [1] "Our mission is to promote sharing and reuse of video data. Most developmental scientists rely on video recordings to capture the complexity and richness of behavior. However, researchers rarely share video data, and this has impeded scientific progress. Enabling reuse of video data is essential to increasing openness, maximizing the efficiency of limited resources, and accelerating the pace of discovery in the developmental and learning sciences.\n\nWe aim to provide tools that encourage good data management, facilitate open data sharing, enable discovery, and foster data reuse and integration. We are working to build functionality that allows researchers to organize, store, code, and share identifiable video data and other contextual information about the videos with other researchers through a secure web-based library.\n\nThe Databrary project is dedicated to transforming the culture of developmental science by building a community of researchers committed to open video data sharing, training a new generation of developmental scientists and empowering them with an unprecedented set of tools for discovery, and raising the profile of behavioral science by bolstering interest in and support for scientific research among the general public."
#> 
#> $doi
#> [1] "10.17910/B7159Q"
#> 
#> $creation
#> [1] "2013-01-11T10:26:40Z"
#> 
#> $owners
#>               name   id
#> 1    Adolph, Karen    5
#> 2 Gilmore, Rick O.    6
#> 3            Staff  308
#> 4 Admin, Databrary 3341
#> 
#> $permission
#> [1] 1
#> 
#> $publicsharefull
#> [1] TRUE
#> 
#> $publicaccess
#> [1] "full"
#>