Lists assets in a given Databrary volume and session (slot).

Lists assets in a given Databrary volume and session (slot).

list_assets_in_session(session_id = 9807, vb = FALSE)

Arguments

session_id

Slot/session ID.

vb

A Boolean value. If TRUE provides verbose output.

Value

A data frame with the assets in the selected volume and session.

Examples

list_assets_in_session()
#>   asset_id asset_type_id                                 name permission
#> 1   153109             2            volumes-citations-monthly          0
#> 2   153108             2 databrary-institutions-investigators          1
#> 3   117035            16         counting (demo Datavyu file)          1
#> 4      314             6               NIH Proposal Narrative          1
#> 5     9200             5                                 logo          1
#> 6        1          -800                counting (demo video)          1
#> 7      315             6               NSF Proposal Narrative          1
#>   classification    size duration session_id                mimetype extension
#> 1             NA      NA       NA       9807                text/csv       csv
#> 2              3    1826       NA       9807                text/csv       csv
#> 3              3    1030       NA       9807 application/vnd.datavyu       opf
#> 4              3 3826718       NA       9807         application/pdf       pdf
#> 5              3  104419       NA       9807               image/png       png
#> 6              3  394499    40000       9807               video/mp4       mp4
#> 7              3 4120930       NA       9807         application/pdf       pdf
#>                  asset_type transcodable
#> 1    Comma-separated values           NA
#> 2    Comma-separated values           NA
#> 3                   Datavyu           NA
#> 4         Portable document           NA
#> 5 Portable network graphics           NA
#> 6              MPEG-4 video         -800
#> 7         Portable document           NA