Lists containers and record structure from Databrary volume.

Lists containers and record structure from Databrary volume.

list_containers_records_json(vol_id = 1, vb = FALSE)

Arguments

vol_id
Databrary volume number.

vb
A Boolean value. If TRUE provides verbose output.

Value

JSON file with containers and records from the specified volume.

Examples

list_containers_records_json()

#> [1] "{\"id\":1,\"name\":\"Databrary sponsored workshops and events\",\"body\":\"Our mission is to promote sharing and reuse of video data. Most developmental scientists rely on video recordings to capture the complexity and richness of behavior. However, researchers rarely share video data, and this has impeded scientific progress. Enabling reuse of video data is essential to increasing openness, maximizing the efficiency of limited resources, and accelerating the pace of discovery in the developmental and learning sciences.\\n\\nWe aim to provide tools that encourage good data management, facilitate open data sharing, enable discovery, and foster data reuse and integration. We are working to build functionality that allows researchers to organize, store, code, and share identifiable video data and other contextual information about the videos with other researchers through a secure web-based library.\\n\\nThe Databrary project is dedicated to transforming the culture of developmental science by building a community of researchers committed to open video data sharing, training a new generation of developmental scientists and empowering them with an unprecedented set of tools for discovery, and raising the profile of behavioral science by bolstering interest in and support for scientific research among the general public.\",\"doi\":\"10.17910/B7159Q\",\"creation\":\"2013-01-11T10:26:40Z\",\"owners\":[{\"name\":\"Adolph, Karen\",\"id\":5},{\"name\":\"Gilmore, Rick O.\",\"id\":6},{\"name\":\"Staff\",\"id\":308},{\"name\":\"Admin, Databrary\",\"id\":3341}],\"permission\":1,\"publicsharefull\":true,\"publicaccess\":\"full\",\"containers\":[{\"id\":1,\"top\":true},{\"id\":6256,\"name\":\"Advisory Board Meeting\",\"date\":\"2013-10-28\",\"release\":3},{\"id\":6257,\"name\":\"Advisory Board Meeting\",\"date\":\"2014-04-07\",\"release\":3},{\"id\":6540,\"name\":\"Advisory Board Meeting\",\"date\":\"2014-10-17\",\"release\":3},{\"id\":7821,\"name\":\"SRCD Preconference Workshop\",\"date\":\"2015-03-18\",\"release\":3},{\"id\":8460,\"name\":\"Advisory Board Meeting\",\"date\":\"2015-05-11\",\"release\":3},{\"id\":9224,\"name\":\"Regional Workshop at NYU (Slides only)\",\"date\":\"2015-09-18\",\"release\":3},{\"id\":9578,\"name\":\"CDS Preconference Workshop\",\"date\":\"2015-10-08\",\"release\":3},{\"id\":9807,\"top\":true,\"name\":\"Top-level materials\"},{\"id\":10068,\"name\":\"Regional Workshop in Philadelphia\",\"date\":\"2015-11-20\",\"release\":3},{\"id\":10991,\"name\":\"Regional Workshop at Penn State\",\"date\":\"2016-03-25\",\"release\":3},{\"id\":11362,\"name\":\"Regional Workshop at University of Maryland, College Park\",\"date\":\"2016-04-15\",\"release\":3},{\"id\":11660,\"name\":\"ICIS16 Preconference Workshop\",\"date\":\"2016-05-25\",\"release\":3},{\"id\":11886,\"name\":\"Advisory Board Meeting\",\"date\":\"2016-06-13\",\"release\":3},{\"id\":14297,\"name\":\"Regional Workshop at University of Minnesota\",\"date\":\"2016-10-14\",\"release\":3},{\"id\":14550,\"name\":\"Regional Workshop at Harvard\",\"date\":\"2016-10-21\",\"release\":3},{\"id\":14551,\"name\":\"Regional Workshop in Boston\",\"date\":\"2016-10-21\",\"release\":3},{\"id\":15663,\"name\":\"Bethesda PLAY Workshop\",\"date\":\"2016-12-16\",\"release\":3},{\"id\":16104,\"name\":\"Regional Workshop at University of Iowa\",\"date\":\"2017-02-03\",\"release\":3},{\"id\":16891,\"name\":\"Pre-Conference at SRCD\",\"date\":\"2017-04-05\",\"release\":3},{\"id\":16894,\"name\":\"Pre-Conference at SRCD \\\"The Science of Play and Learning\\\"\",\"date\":\"2017-04-05\",\"release\":3},{\"id\":17187,\"name\":\"AERA 2017 San Antonio\",\"date\":\"2017-04-30\",\"release\":3},{\"id\":17417,\"name\":\"Regional Workshop at Emory University in Atlanta, GA\",\"date\":\"2017-05-15\",\"release\":3},{\"id\":17886,\"name\":\"Advisory Board Meeting 2017\",\"date\":\"2017-06-09\",\"release\":3},{\"id\":21077,\"name\":\"CDS 2017 Portland\",\"date\":\"2017-10-13\",\"release\":3},{\"id\":25952,\"name\":\"Advisory Board Meeting 2018\",\"date\":\"2018-03-26\",\"release\":3}],\"records\":[{\"id\":5481,\"category\":6,\"measures\":{\"29\":\"Datavyu & Best Practices for Coding Video\"}},{\"id\":5480,\"category\":6,\"measures\":{\"29\":\"Managing Data with Databrary\"}},{\"id\":5479,\"category\":6,\"measures\":{\"29\":\"Policies for Sharing Identifiable Video\"}},{\"id\":5478,\"category\":6,\"measures\":{\"29\":\"Introduction to Databrary\"}}]}"